requirements.txt
pytest~=9.0.1
selenium~=4.38.0
selene~=2.0.0rc9
allure-pytest==2.15.0
allure-python-commons~=2.15.0
selenium~=4.38.0
utils~=1.0.2
python-dotenv~=1.2.1